+++ UnitTests/BraintreePayPalTests/BTPayPalClient_Tests.swift
@@ -19,7 +19,7 @@ class BTPayPalClient_Tests: XCTestCase {
         mockAPIClient.cannedResponseBody = BTJSON(value: [
             "paymentResource": ["redirectUrl": "http://fakeURL.com"]
         ])
-        payPalClient = BTPayPalClient(apiClient: mockAPIClient)
+        payPalClient = BTPayPalClient(apiClient: mockAPIClient, universalLink: URL(string: "https://www.paypal.com")!)
         mockWebAuthenticationSession = MockWebAuthenticationSession()
         payPalClient.webAuthenticationSession = mockWebAuthenticationSession
     }
@@ -237,8 +237,7 @@ class BTPayPalClient_Tests: XCTestCase {
 
         let vaultRequest = BTPayPalVaultRequest(
             userAuthenticationEmail: "fake@gmail.com",
-            enablePayPalAppSwitch: true,
-            universalLink: URL(string: "https://www.paypal.com")!
+            enablePayPalAppSwitch: true
         )
 
         mockAPIClient.cannedResponseBody = BTJSON(value: [
@@ -749,8 +748,7 @@ class BTPayPalClient_Tests: XCTestCase {
         
         let vaultRequest = BTPayPalVaultRequest(
             userAuthenticationEmail: "fake@gmail.com",
-            enablePayPalAppSwitch: true,
-            universalLink: URL(string: "https://paypal.com")!
+            enablePayPalAppSwitch: true
         )
 
         payPalClient.tokenize(vaultRequest) { _, _ in }
@@ -785,8 +783,7 @@ class BTPayPalClient_Tests: XCTestCase {
 
         let vaultRequest = BTPayPalVaultRequest(
             userAuthenticationEmail: "fake@gmail.com",
-            enablePayPalAppSwitch: true,
-            universalLink: URL(string: "https://paypal.com")!
+            enablePayPalAppSwitch: true
         )
 
         let expectation = expectation(description: "completion block called")
@@ -810,8 +807,7 @@ class BTPayPalClient_Tests: XCTestCase {
 
         let vaultRequest = BTPayPalVaultRequest(
             userAuthenticationEmail: "fake@gmail.com",
-            enablePayPalAppSwitch: true,
-            universalLink: URL(string: "https://paypal.com")!
+            enablePayPalAppSwitch: true
         )
 
         mockAPIClient.cannedResponseBody = BTJSON(value: [
@@ -834,8 +830,7 @@ class BTPayPalClient_Tests: XCTestCase {
     func testHandleReturn_whenURLIsCancel_returnsCancel() {
         let request = BTPayPalVaultRequest(
             userAuthenticationEmail: "sally@gmail.com",
-            enablePayPalAppSwitch: true,
-            universalLink: URL(string: "https://merchant-app.com/merchant-path")!
+            enablePayPalAppSwitch: true
         )
         let returnURL = URL(string: "https://www.merchant-app.com/merchant-path/cancel?ba_token=A_FAKE_BA_TOKEN&switch_initiated_time=1234567890")!
         let expectation = expectation(description: "completion block called")
@@ -856,8 +851,7 @@ class BTPayPalClient_Tests: XCTestCase {
     func testHandleReturn_whenURLIsUnknown_returnsError() {
         let request = BTPayPalVaultRequest(
             userAuthenticationEmail: "sally@gmail.com",
-            enablePayPalAppSwitch: true,
-            universalLink: URL(string: "https://merchant-app.com/merchant-path")!
+            enablePayPalAppSwitch: true
         )
         let returnURL = URL(string: "https://www.merchant-app.com/merchant-path/garbage-url")!
         let expectation = expectation(description: "completion block called")
@@ -878,8 +872,7 @@ class BTPayPalClient_Tests: XCTestCase {
     func testHandleReturn_whenURLIsSuccess_returnsTokenization() {
         let request = BTPayPalVaultRequest(
             userAuthenticationEmail: "sally@gmail.com",
-            enablePayPalAppSwitch: true,
-            universalLink: URL(string: "https://merchant-app.com/merchant-path")!
+            enablePayPalAppSwitch: true
         )
         mockAPIClient.cannedResponseBody = BTJSON(value: [
             "paypalAccounts":
@@ -931,8 +924,7 @@ class BTPayPalClient_Tests: XCTestCase {
 
         let vaultRequest = BTPayPalVaultRequest(
             userAuthenticationEmail: "fake@gmail.com",
-            enablePayPalAppSwitch: true,
-            universalLink: URL(string: "https://paypal.com")!
+            enablePayPalAppSwitch: true
         )
 
         mockAPIClient.cannedResponseBody = BTJSON(value: [
@@ -949,7 +941,7 @@ class BTPayPalClient_Tests: XCTestCase {
         XCTAssertEqual(lastPostParameters["launch_paypal_app"] as? Bool, true)
         XCTAssertTrue((lastPostParameters["os_version"] as! String).matches("\\d+\\.\\d+"))
         XCTAssertTrue((lastPostParameters["os_type"] as! String).matches("iOS|iPadOS"))
-        XCTAssertEqual(lastPostParameters["merchant_app_return_url"] as? String, "https://paypal.com")
+        XCTAssertEqual(lastPostParameters["merchant_app_return_url"] as? String, "https://www.paypal.com")
     }
 
     func testIsiOSAppSwitchAvailable_whenApplicationCantOpenPayPalInAppURL_returnsFalseAndSendsAnalytics() {
@@ -959,8 +951,7 @@ class BTPayPalClient_Tests: XCTestCase {
 
         let vaultRequest = BTPayPalVaultRequest(
             userAuthenticationEmail: "fake@gmail.com",
-            enablePayPalAppSwitch: true,
-            universalLink: URL(string: "https://paypal.com")!
+            enablePayPalAppSwitch: true
         )
 
         mockAPIClient.cannedResponseBody = BTJSON(value: [